+++ src/extension.ts
@@ -1,7 +1,24 @@
 import * as process from 'child_process';
 import * as vscode from 'vscode';
+import { DOMParser } from 'xmldom';
+
+import { Decorator } from './decoration';
 
 export function activate(context: vscode.ExtensionContext) {
+
+	if (vscode.workspace.workspaceFolders === undefined) { return; }
+	if (vscode.workspace.workspaceFolders.length > 1) {
+		vscode.window.showWarningMessage(`TFSC Extension don't works with multy-root workspace!`);
+		return;
+	}
+
+	const decorator = new Decorator(context);
+
+	const rootUri = vscode.workspace.workspaceFolders?.[0]?.uri;
+	const scm = vscode.scm.createSourceControl('tfvc', "TF Version Control", rootUri);
+	const changes = scm.createResourceGroup('changes', 'Changes');
+	refreshSCM(scm, changes, decorator);
+
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-tfvc.checkoutCurrentFile', () => {
@@ -11,6 +28,7 @@ export function activate(context: vscode.ExtensionContext) {
 				return;
 			}
 			checkout(fileName);
+			vscode.commands.executeCommand('vscode-tfvc.refreshSCM');
 		}),
 		vscode.commands.registerCommand('vscode-tfvc.checkInCurrentFile', () => {
 			const fileName = vscode.window.activeTextEditor?.document.fileName;
@@ -19,6 +37,7 @@ export function activate(context: vscode.ExtensionContext) {
 				return;
 			}
 			checkIn(fileName);
+			vscode.commands.executeCommand('vscode-tfvc.refreshSCM');
 		}),
 		vscode.commands.registerCommand('vscode-tfvc.undoCurrentFile', () => {
 			const fileName = vscode.window.activeTextEditor?.document.fileName;
@@ -27,11 +46,63 @@ export function activate(context: vscode.ExtensionContext) {
 				return;
 			}
 			undo(fileName);
+			vscode.commands.executeCommand('vscode-tfvc.refreshSCM');
+		}),
+		vscode.commands.registerCommand('vscode-tfvc.refreshSCM', () => {
+			refreshSCM(scm, changes, decorator);
+		}),
+
+		vscode.commands.registerCommand('vscode-tfvc.undoFile', (state: vscode.SourceControlResourceState) => {
+			undo(state.resourceUri.fsPath);
+		}),
+		vscode.commands.registerCommand('vscode-tfvc.checkInFile', (state: vscode.SourceControlResourceState) => {
+			checkIn(state.resourceUri.fsPath);
+		}),
+		vscode.commands.registerCommand('vscode-tfvc.openFile', (state: vscode.SourceControlResourceState) => {
+			vscode.commands.executeCommand('vscode.open', state.resourceUri);
 		}),
 
 		vscode.workspace.onWillSaveTextDocument(saveEvent => {
 			const fileName = saveEvent.document.fileName;
 			saveEvent.waitUntil(checkout(fileName));
+		}),
+
+		vscode.workspace.onDidCreateFiles(() => vscode.commands.executeCommand('vscode-tfvc.refreshSCM')),
+		vscode.workspace.onDidDeleteFiles(() => vscode.commands.executeCommand('vscode-tfvc.refreshSCM')),
+		vscode.workspace.onDidRenameFiles(() => vscode.commands.executeCommand('vscode-tfvc.refreshSCM')),
+		vscode.workspace.onDidSaveTextDocument(() => vscode.commands.executeCommand('vscode-tfvc.refreshSCM')),
+	);
+}
+
+function refreshSCM(scm: vscode.SourceControl, group: vscode.SourceControlResourceGroup, decorator: Decorator) {
+	const configuration = vscode.workspace.getConfiguration('vscode-tfvc');
+	const tfPath = configuration.get<string>('tfExePath');
+	// todo: check if exist and if not ask
+	if (tfPath === undefined || tfPath.length === 0) { return Promise.reject(); }
+
+	return vscode.window.withProgress(
+		{ title: `Refreshing the source control...`, location: vscode.ProgressLocation.SourceControl },
+		_progress => new Promise((resolve, reject) => {
+			const childProcess = process.execFile(tfPath, ['vc', 'status', scm.rootUri?.fsPath ?? '', '/recursive', '/format:xml']);
+			let message = '', error = '';
+			childProcess.stdout?.on('data', data => message += data.toString());
+			childProcess.stderr?.on('data', data => error += data.toString());
+			childProcess.on("exit", code => {
+				if (code === 0) {
+					const parser = new DOMParser();
+					const xmlDoc = parser.parseFromString(message, "text/xml");
+					const elements = Array.from(xmlDoc.getElementsByTagName('PendingChange'));
+					group.resourceStates = elements.map(el => ({
+						resourceUri: vscode.Uri.file(el.getAttribute('local')!),
+						decorations: decorator.getDecorations(el.getAttribute('chg') ?? ''),
+					}));
+					resolve();
+				}
+				else {
+					vscode.window.showErrorMessage(`Error: Cannot check the source control! (Code: ${code}; Error: ${error})`);
+					reject();
+				}
+			});
 		})
 	);
 }