+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to the "vscode-tfvc" extension will be documented in this fi
 
 ## [Unreleased]
 
+## [0.3.0] 2020-07-15
+### Added
+- Basic `SCM Side Bar`. Include:
+  - Display the changed files and their status (`Added`, `Deleted`, `Edited` or `Renamed`).
+  - Allow to check-in, undo & open files.
+  - Manual refresh (via `refreshSCM`) and auto refresh on changes and after TFVC commands.
+
+
+## [0.2.0] 2020-07-06
+### Added
+- `Check in the current file` command.
+- `Undo the pending changes in the current file` command.
+
 ## [0.1.1] 2020-07-05
 ### fixed
 - Saving a file will wait until the checkout is over.
@@ -14,6 +27,6 @@ All notable changes to the "vscode-tfvc" extension will be documented in this fi
 ## [0.1.0] 2020-07-05
 
 ### Added
-- `Check out current file` command.
+- `Check out the current file` command.
 - Automatic checking out on save.
 - `vscode-tfvc.checkoutCurrentFile` configuration.